fix(plugin-meetings): addressed review comments for new class reachabilityPeerConnection
